requirements.txt
fastapi
uvicorn
jinja2<3.1.5
apscheduler
supabase
pandas
numpy
hmmlearn
scikit-learn
joblib
openpyxl
xlrd
xgboost
yfinance
curl_cffi
python-dotenv
optuna
scipy
shap
catboost
groq
arch
xmltodict
httpx
